# docs/openapi/connector-v1.yaml
openapi: 3.1.0
info:
  title: API внешних коннекторов Scheduler
  version: 1.0.0
  description: API для приёма полных снимков расписания с проверкой подписи.
servers:
  - url: https://admin.example.org
paths:
  /api/v1/connector-spec:
    get:
      summary: Получить действующий контракт коннектора
      responses:
        "200":
          description: Сведения о контракте
  /api/v1/connectors/{connectorId}/snapshots:
    post:
      summary: Отправить полный снимок расписания
      security: [{ ed25519: [] }]
      parameters:
        - $ref: "#/components/parameters/ConnectorId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schema/schedule-snapshot-v1.json"
      responses:
        "202": { description: Снимок поставлен в очередь }
        "200": { description: Повтор ранее принятого снимка }
        "401": { description: Подпись отклонена }
        "413": { description: Превышен допустимый размер данных }
        "422": { description: Данные не соответствуют контракту }
        "429": { description: Превышено ограничение частоты запросов }
  /api/v1/connectors/{connectorId}/runs/{runId}:
    get:
      summary: Получить состояние обработки
      security: [{ ed25519: [] }]
      parameters:
        - $ref: "#/components/parameters/ConnectorId"
        - name: runId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Текущее состояние обработки
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunStatus"
        "404": { description: Запуск не найден }
components:
  schemas:
    RunStatus:
      type: object
      required: [run_id, status, ingestion_sequence, received_at]
      properties:
        run_id: { type: string }
        status:
          type: string
          enum:
            [
              received,
              processing,
              staged,
              quarantined,
              published,
              rejected,
              failed,
              superseded,
            ]
        ingestion_sequence:
          type: integer
          minimum: 1
          description: Монотонный номер приёма сервером в рамках источника. Не меняется при повторе idempotency key.
        received_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
  securitySchemes:
    ed25519:
      type: apiKey
      in: header
      name: X-Scheduler-Signature
      description: Правила формирования подписи описаны в docs/connector-api.md.
  parameters:
    ConnectorId:
      name: connectorId
      in: path
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 200 }